+++ CHANGELOG.rst
@@ -10,6 +10,9 @@ Unreleased
 Changed
 ~~~~~~~
 
+* http-client: Remove deprecated method `get` from c.t.finatra.httpclient.HttpClient.
+  Use HttpClient's `execute` instead. ``PHAB_ID=D618904``
+
 * finatra: Create the finatra/http-core project, which is meant to contain common artifacts
            for the finatra/http-server and finatra/http-client project. As part of this
            change, the `com.twitter.finatra.httpclient.RequestBuilder` has been deprecated